Rename calendar picker folder name
