+++ src/app/events/event-create/calendar-picker/calendar-picker.component.scss
